@@ -11,24 +11,27 @@
     ],
     "license": "CC-BY-4.0",
     "scripts": {
-        "build:png": "node ./scripts/processPNG.js",
-        "build:zip": "node ./scripts/processZIP.js",
+        "build:png": "ts-node ./scripts/processPNG.ts",
+        "build:zip": "ts-node ./scripts/processZIP.ts",
         "type-check": "tsc --noEmit",
         "lint": "eslint --ignore-path .gitignore . --ext ts --ext tsx --ext js --ext jsx",
         "lint:fix": "yarn lint --fix"
     },
     "devDependencies": {
-        "@typescript-eslint/eslint-plugin": "^1.9.0",
-        "@typescript-eslint/parser": "^1.9.0",
-        "eslint": "^5.16.0",
-        "eslint-config-prettier": "^4.3.0",
+        "@types/node": "^12.7.2",
+        "@typescript-eslint/eslint-plugin": "^2.0.0",
+        "@typescript-eslint/parser": "^2.0.0",
+        "archiver": "^3.1.1",
+        "eslint": "^6.2.0",
+        "eslint-config-prettier": "^6.1.0",
         "eslint-plugin-prettier": "^3.1.0",
-        "eslint-plugin-react": "^7.13.0",
+        "eslint-plugin-react": "^7.14.3",
         "glob": "^7.1.4",
         "ora": "^3.4.0",
-        "prettier": "^1.17.1",
-        "sharp": "^0.22.1",
-        "typescript": "^3.4.5"
+        "prettier": "^1.18.2",
+        "sharp": "^0.23.0",
+        "ts-node": "^8.3.0",
+        "typescript": "^3.5.3"
     },
     "prettier": {
         "singleQuote": false,
